Finish and add getAudio route
