+++ backend/src/index.ts
@@ -6,11 +6,12 @@ import * as Tracing from "@sentry/tracing";
 import { checkQuota, UserRequest, validateAuth } from "./helpers.js";
 import { changePDPpassphrase, disablePDP, enablePDP } from "./pdp.js";
 import { deleteAccount, getOrCreateKeys, sync } from "./accounts.js";
-import { getImage, moodLog, survey } from "./main.js";
+import { moodLog, survey } from "./main.js";
 import { gapFund } from "./gap.js";
 import { beacon } from "./analytics.js";
 import { graniteLink } from "./granite.js";
 import { search } from "./spotify.js";
+import { getAudio, getImage } from "./storage.js";
 
 const app = express();
 initializeApp({
@@ -67,6 +68,7 @@ app.post("/analytics/beacon", beacon);
 app.post("/moodLog", moodLog);
 app.post("/survey", survey);
 app.post("/getImage", getImage);
+app.post("/getAudio", getAudio);
 
 app.post("/gap", gapFund);
 app.post("/spotify/search", search);
