+++ backend/src/main.ts
@@ -13,45 +13,6 @@ import { PubSub } from "@google-cloud/pubsub";
 
 const pubsub = new PubSub();
 
-export const getImage = async (req: UserRequest, res: Response) => {
-    const db = getDatabase();
-
-    const body = req.body;
-    const encryptionKey = await validateKeys(body.keys, db, req.user!.user_id);
-    if (!encryptionKey) {
-        res.send(400);
-        return;
-    }
-
-    const filename = body.filename;
-    // Max length: 32-char UUID.webp.enc
-    if (typeof filename !== "string" || filename.length > 45) {
-        res.send(400);
-        return;
-    }
-
-    let file;
-    try {
-        file = await getStorage().bucket().file(`user/${req.user!.uid}/${filename}`).download();
-    } catch {
-        // File doesn't exist, fail
-        res.send(400);
-        return;
-    }
-
-    // Null check for type safety
-    if (!file) {
-        res.send(400);
-        return;
-    }
-
-    if (filename.endsWith(".enc")) {
-        res.send(AES.decrypt(file[0].toString("utf8"), encryptionKey).toString(aesutf8));
-    } else {
-        res.send(`data:image/webp;base64,${file[0].toString("base64")}`);
-    }
-}
-
 export const survey = async (req: UserRequest, res: Response) => {
     const body = req.body;
     const db = getDatabase();
